Fixed marker-mid attribute of svg <path> element.
I replaced the current PathParser.java with SVGPathParser (the old one is still there
but it has no function anymore. Should be removed if this pull request is accepted)
The new SVGPathParser remembers all marker points. These can be used later in the
MarkerBuilder.
The parsed data is remembered for every SVGPath by adding it to it's properties.